% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nickname_dictionary.R
\name{get_nickname_dictionary}
\alias{get_nickname_dictionary}
\title{Cached access to the nickname dictionary}
\usage{
get_nickname_dictionary(refresh = FALSE)
}
\arguments{
\item{refresh}{rebuild even if cached.}
}
\value{
see [create_nickname_dictionary()].
}
\description{
Cached access to the nickname dictionary
}
