% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_person.R
\docType{data}
\name{NAME_NOISE}
\alias{NAME_NOISE}
\title{Credential and title tokens seen in provider directories.}
\format{
An object of class \code{character} of length 50.
}
\usage{
NAME_NOISE
}
\description{
Stripped BEFORE parsing: a name parser has no way to know `CRNP` is not a
middle name.
}
\keyword{datasets}
